+++ lib/xnix/compatibility/dbus_runtime_client.rb
@@ -65,6 +65,10 @@ module Xnix
         parse_dictionary(call("GetDesktopEntryPlan", application_id))
       end
 
+      def task_manager_identity_plan(application_id)
+        parse_dictionary(call("GetTaskManagerIdentityPlan", application_id))
+      end
+
       def state_root(application_id)
         parse_dictionary(call("GetApplicationStateRoot", application_id))
       end